refactor(sync): extract sync/transactions.rs
Move the read-only query surface out of sync/mod.rs into a dedicated
transactions submodule. Everything in the new file is a "ask the
wallet a question" helper that the FRB layer in api/sync.rs or the
C FFI layer in ffi.rs calls per user action:

  - Balance: WalletBalance + get_wallet_balance
  - Address: get_next_available_address
  - Enhancement queue: TxDataRequest + get_transaction_data_requests
                       + decrypt_and_store_transaction
                       + set_transaction_status
  - History: TransactionInfo + get_transaction_history
  - Pending TX tracking: PendingTxInfo + get_pending_transactions
                         + check_tx_mined

None of these belong to the orchestration loop (that lives in
sync_engine). Pulling them into their own submodule finishes what
the refactor started — sync/mod.rs is now a ~270 line "glue" module
that owns the per-wallet infrastructure (DB open, chain-tip update,
scan range management, rewind, validate_address, get_blocks_dir) and
the shared PROPOSAL_STORE used by both send.rs and pczt.rs.

All extracted functions are re-exported via `pub(crate) use
transactions::{...}` so every `crate::wallet::sync::get_wallet_balance`
caller in api/sync.rs and ffi.rs resolves unchanged. The struct
types (`WalletBalance`, `TransactionInfo`, `TxDataRequest`,
`PendingTxInfo`) are NOT re-exported: the FRB layer has its own
mirror structs and only touches fields via `.transparent`,
`.txid_hex`, etc., so nothing outside sync/transactions.rs names
them.

23 tests pass.
